# supabase/schema-openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "standard public schema",
    "version": "14.1"
  },
  "host": "wjcbobcqyqdkludsbqgf.supabase.co:443",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "produces": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "paths": {
...
    "/avaliacoes": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.id"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.nome"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.tipo"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.descricao"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.disciplinaid"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.turmaid"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.datarealizacao"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.peso"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.pontuacaomaxima"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.dataatualizacao"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.numero_id"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/avaliacoes"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "avaliacoes"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.avaliacoes"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "avaliacoes"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.id"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.nome"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.tipo"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.descricao"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.disciplinaid"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.turmaid"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.datarealizacao"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.peso"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.pontuacaomaxima"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.dataatualizacao"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.numero_id"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "avaliacoes"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.id"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.nome"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.tipo"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.descricao"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.disciplinaid"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.turmaid"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.datarealizacao"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.peso"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.pontuacaomaxima"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.dataatualizacao"
          },
          {
            "$ref": "#/parameters/rowFilter.avaliacoes.numero_id"
          },
          {
            "$ref": "#/parameters/body.avaliacoes"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "avaliacoes"
        ]
      }
    },
...
  },
  "definitions": {
...
    "avaliacoes": {
      "required": [
        "id",
        "nome",
        "numero_id"
      ],
      "properties": {
        "id": {
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "integer",
          "type": "integer"
        },
        "nome": {
          "format": "character varying",
          "maxLength": 255,
          "type": "string"
        },
        "tipo": {
          "format": "character varying",
          "maxLength": 50,
          "type": "string"
        },
        "descricao": {
          "format": "text",
          "type": "string"
        },
        "disciplinaid": {
          "description": "Note:\nThis is a Foreign Key to `disciplinas.id`.<fk table='disciplinas' column='id'/>",
          "format": "integer",
          "type": "integer"
        },
        "turmaid": {
          "description": "Note:\nThis is a Foreign Key to `turmas.id`.<fk table='turmas' column='id'/>",
          "format": "integer",
          "type": "integer"
        },
        "datarealizacao": {
          "format": "date",
          "type": "string"
        },
        "peso": {
          "default": 1,
          "format": "numeric",
          "type": "number"
        },
        "pontuacaomaxima": {
          "default": 10,
          "format": "numeric",
          "type": "number"
        },
        "datacriacao": {
          "default": "CURRENT_TIMESTAMP",
          "format": "timestamp without time zone",
          "type": "string"
        },
        "dataatualizacao": {
          "default": "CURRENT_TIMESTAMP",
          "format": "timestamp without time zone",
          "type": "string"
        },
        "numero_id": {
          "format": "integer",
          "type": "integer"
        }
      },
      "type": "object"
    },
...
  },
  "parameters": {
...
    "preferReturn": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferCount": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "count=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferPost": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none",
        "resolution=ignore-duplicates",
        "resolution=merge-duplicates"
      ],
      "in": "header",
      "type": "string"
    },
    "select": {
      "name": "select",
      "description": "Filtering Columns",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "order": {
      "name": "order",
      "description": "Ordering",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "range": {
      "name": "Range",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "header",
      "type": "string"
    },
    "rangeUnit": {
      "name": "Range-Unit",
      "description": "Limiting and Pagination",
      "required": false,
      "default": "items",
      "in": "header",
      "type": "string"
    },
    "offset": {
      "name": "offset",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "limit": {
      "name": "limit",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.avaliacoes": {
      "name": "avaliacoes",
      "description": "avaliacoes",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/avaliacoes"
      }
    },
    "rowFilter.avaliacoes.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.avaliacoes.nome": {
      "name": "nome",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.avaliacoes.tipo": {
      "name": "tipo",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.avaliacoes.descricao": {
      "name": "descricao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.avaliacoes.disciplinaid": {
      "name": "disciplinaid",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.avaliacoes.turmaid": {
      "name": "turmaid",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.avaliacoes.datarealizacao": {
      "name": "datarealizacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.avaliacoes.peso": {
      "name": "peso",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.avaliacoes.pontuacaomaxima": {
      "name": "pontuacaomaxima",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.avaliacoes.datacriacao": {
      "name": "datacriacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.avaliacoes.dataatualizacao": {
      "name": "dataatualizacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.avaliacoes.numero_id": {
      "name": "numero_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
  },
...
}